MIME4J-262 bugfix of MessageBuilder regarding getDate
Before MessageBuilder produces message that had a getDate that was returning a
non null date instead of null when the message had no "Date" header.

In order to not break old code that do not check for null value, old way to
create a builder has been kept and depreciated. Those old way return a builder
that act exactly the same than before.

New way of building a MessageBuilder has been added and do create a builder that
are not buggy anymore.